requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
more-itertools==4.2.0
mysqlclient==1.3.12
oauthlib==2.1.0
openpyxl==2.5.3
parso==0.2.1
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.6.0
prompt-toolkit==1.0.15
py==1.5.3
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.8.1
pytest==3.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.19.0
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
whitenoise==3.3.1
